refactor(client): move inventory slice into features/inventory
