+++ client/src/features/inventory/components/AdjustStockDialog.tsx
@@ -6,13 +6,19 @@ import {
   DialogDescription,
   DialogHeader,
   DialogTitle,
-} from '../shared/ui/dialog';
-import { Button } from '../shared/ui/button';
-import { Input } from '../shared/ui/input';
-import { Label } from '../shared/ui/label';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../shared/ui/select';
-import { resource } from '../shared/lib/resource';
-import { useTranslation } from '../shared/i18n/index';
+} from '../../../shared/ui/dialog';
+import { Button } from '../../../shared/ui/button';
+import { Input } from '../../../shared/ui/input';
+import { Label } from '../../../shared/ui/label';
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from '../../../shared/ui/select';
+import { resource } from '../../../shared/lib/resource';
+import { useTranslation } from '../../../shared/i18n/index';
 import type { Product } from '@/types';
 
 const products = resource<Product>('products');
