+++ client/src/features/inventory/components/inventory/BulkOperationDialogs.tsx
@@ -1,6 +1,6 @@
-import { Button } from '../../shared/ui/button';
-import { Input } from '../../shared/ui/input';
-import { Label } from '../../shared/ui/label';
+import { Button } from '../../../../shared/ui/button';
+import { Input } from '../../../../shared/ui/input';
+import { Label } from '../../../../shared/ui/label';
 import {
   Dialog,
   DialogContent,
@@ -8,7 +8,7 @@ import {
   DialogTitle,
   DialogFooter,
   DialogDescription,
-} from '../../shared/ui/dialog';
+} from '../../../../shared/ui/dialog';
 import {
   AlertDialog,
   AlertDialogAction,
@@ -18,15 +18,15 @@ import {
   AlertDialogFooter,
   AlertDialogHeader,
   AlertDialogTitle,
-} from '../../shared/ui/alert-dialog';
+} from '../../../../shared/ui/alert-dialog';
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '../../shared/ui/select';
-import { useTranslation } from '../../shared/i18n/index';
+} from '../../../../shared/ui/select';
+import { useTranslation } from '../../../../shared/i18n/index';
 import type { Category, Distributor } from '@/types';
 
 interface BulkOperationDialogsProps {
