+++ client/src/pages/BarcodeTools.tsx
@@ -9,7 +9,7 @@ import { Input } from '../shared/ui/input';
 import { Badge } from '../shared/ui/badge';
 import { Checkbox } from '../shared/ui/checkbox';
 import BarcodeScanner from '../shared/components/BarcodeScanner';
-import BarcodeGenerator from '../components/BarcodeGenerator';
+import BarcodeGenerator from '../features/inventory/components/BarcodeGenerator';
 import { formatCurrency } from '../shared/lib/utils';
 import { useCartStore } from '../store/cartStore';
 import { useApiQuery } from '../shared/lib/apiQuery';
